+++ demos/manual/src/main/java/is/codion/manual/store/Misc.java
@@ -61,7 +61,7 @@ public final class Misc {
 		JasperReportsDataSource<Entity> dataSource =
 						new JasperReportsDataSource<>(customerIterator,
 										(entity, reportField) ->
-														entity.get(customerDefinition.attributes().get(reportField.getName())));
+														entity.get(customerDefinition.attributes().getOrThrow(reportField.getName())));
 
 		JRReport customerReport = fileReport("reports/customer.jasper");