+++ framework/json-domain/src/main/java/is/codion/framework/json/domain/CustomConditionDeserializer.java
@@ -50,7 +50,7 @@ final class CustomConditionDeserializer implements Serializable {
 		String conditionTypeName = conditionNode.get("conditionType").asText();
 		JsonNode columnsNode = conditionNode.get("columns");
 		List<Column<?>> columns = Arrays.stream(entityObjectMapper.readValue(columnsNode.toString(), String[].class))
-						.map(definition.attributes()::get)
+						.map(definition.attributes()::getOrThrow)
 						.map(attribute -> (Column<?>) attribute)
 						.collect(toList());
 		JsonNode valuesNode = conditionNode.get("values");